% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{diffpeak}
\alias{diffpeak}
\title{Differential Peak Analysis}
\usage{
diffpeak(
  summary_list = NULL,
  condition = NULL,
  log2 = TRUE,
  num = NULL,
  den = NULL
)
}
\arguments{
\item{summary_list}{Output from \code{\link{extract_summary}}}

\item{condition}{a column name in \code{coldata} containing sample conditions. Default NULL.}

\item{log2}{log2 convert data prior to testing. Default TRUE}

\item{num}{Numerator condition. Default NULL}

\item{den}{Denominator condition. Default NULL}
}
\description{
Differential Peak Analysis
}
\details{
Takes output from \code{extract_summary} and performs differential peak analysis with Limma
}
