% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{extract_summary}
\alias{extract_summary}
\title{Extract area under the curve for every peak from from given bigWig files.}
\usage{
extract_summary(
  colData,
  bed = NULL,
  ucsc_assembly = TRUE,
  startFrom = "start",
  binSize = 50,
  up = 2500,
  down = 2500,
  pc_genes = TRUE,
  nthreads = 4
)
}
\arguments{
\item{colData}{bigWig files. Default NULL. Required.}

\item{bed}{bed file or a data.frame with first 3 column containing chromosome, star, end positions.}

\item{ucsc_assembly}{If \code{bed} file not provided, setting \code{ucsc_assembly} to TRUE will fetch transcripts from UCSC genome browser.}

\item{startFrom}{Default "start". For bed files this can be "start", "center" or "end". For \code{ucsc_assembly} this can only be "start" or "end"}

\item{binSize}{bin size to extract signal. Default 50 (bps). Should be >1}

\item{up}{extend upstream by this many bps from \code{startFrom}. Default 2500}

\item{down}{extend downstream by this many bps from \code{startFrom}. Default 2500}

\item{pc_genes}{Use only protein coding genes when using \code{ucsc_assembly}. Default TRUE}

\item{nthreads}{Default 4}
}
\description{
Extract area under the curve for every peak from from given bigWig files.
}
