% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{pca_plot}
\alias{pca_plot}
\title{Draw a PCA plot}
\usage{
pca_plot(
  summary_list = NULL,
  top = 1000,
  log2 = FALSE,
  xpc = "PC1",
  ypc = "PC2",
  color_by = NULL,
  pch_by = NULL,
  color = NULL,
  show_cree = TRUE,
  lab_size = 1,
  size = 1,
  legend_size = 1,
  legendpos = "topright",
  legendpos2 = "bottomright"
)
}
\arguments{
\item{summary_list}{output from extract_summary}

\item{top}{Top most variable peaks to consider for PCA. Default 1000}

\item{log2}{log transform data? Default FALSE. IF TRUE, adds a small positive value and log2 converts.}

\item{xpc}{Default PC1}

\item{ypc}{Default PC2}

\item{color_by}{a column name in \code{coldata} to color by. Default NULL.}

\item{pch_by}{a column name in \code{coldata} to pch by. Default NULL.}

\item{color}{Manual colors for each level in \code{color_by} Default NULL.}

\item{show_cree}{If TRUE draws a cree plot. Default TRUE}

\item{lab_size}{Font size for labels. Default 1}

\item{size}{Point size. Default 1}

\item{legend_size}{Default 1}

\item{legendpos}{Default topright}

\item{legendpos2}{Default bottomright}
}
\description{
Draw a PCA plot
}
