% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Draw a heatmap}
\usage{
plot_heatmap(
  mat_list,
  sortBy = "mean",
  col_pal = "Blues",
  revpal = FALSE,
  sample_names = NULL,
  title_size = 1,
  top_profile = FALSE,
  zmins = NULL,
  zmaxs = NULL,
  scale = FALSE,
  file_name = NULL,
  hm_width = 1024,
  hm_height = 600,
  ...
)
}
\arguments{
\item{mat_list}{Input matrix list generated by \code{\link{profile_extract}}}

\item{sortBy}{Sort matrix by.., Can be mean, median. Default mean.}

\item{col_pal}{Color palette to use. Default Blues. Use hcl.pals(type = "sequential") to see available palettes}

\item{revpal}{Reverse color palette? Default FALSE.}

\item{title_size}{size of title. Default 0.8}

\item{top_profile}{Boolean. Whether to draw top profile plot.}

\item{zmins}{Manually specify min scores to include in heatmap}

\item{zmaxs}{Manually specify max scores to include in heatmap}

\item{scale}{Whether to row scale the matrix. Default FALSE}

\item{file_name}{Default NULL. If provided saves plot as a png.}

\item{hm_width}{Width of the plot. Default 1024}

\item{hm_height}{Height of the plot. Default 600}

\item{sample_bames}{Manually specify sample names.}
}
\description{
Draw a heatmap
}
\details{
This function takes output from extract_matrix and draws a heatmap
}
