% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{profile_extract}
\alias{profile_extract}
\title{Generate bigWig signal matrix for given genomic regions or ucsc refseq transcripts}
\usage{
profile_extract(
  colData = NULL,
  bed = NULL,
  ucsc_assembly = TRUE,
  startFrom = "start",
  binSize = 50,
  up = 2500,
  down = 2500,
  pc_genes = TRUE,
  nthreads = 4
)
}
\arguments{
\item{bed}{bed file or a data.frame with first 3 column containing chromosome, star, end positions.}

\item{ucsc_assembly}{If \code{bed} file not provided, setting \code{ucsc_assembly} to TRUE will fetch transcripts from UCSC genome browser.}

\item{startFrom}{Default "start". For bed files this can be "start", "center" or "end". For \code{ucsc_assembly} this can only be "start" or "end"}

\item{binSize}{bin size to extract signal. Default 50 (bps). Should be >1}

\item{up}{extend upstream by this many bps from \code{startFrom}. Default 2500}

\item{down}{extend downstream by this many bps from \code{startFrom}. Default 2500}

\item{pc_genes}{Use only protein coding genes when \code{ucsc_assembly} is used. Default TRUE}

\item{nthreads}{Default 4}

\item{coldata}{from \code{read_coldata}}
}
\description{
Generate bigWig signal matrix for given genomic regions or ucsc refseq transcripts
}
\seealso{
\code{\link{profile_summarize}} \code{\link{profile_plot}} \code{\link{plot_heatmap}}
}
