% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{profile_plot}
\alias{profile_plot}
\title{Draw a profile plot}
\usage{
profile_plot(
  sig_list = NULL,
  color = NULL,
  condition = NULL,
  condition_colors = NULL,
  collapse_replicates = FALSE,
  plot_se = FALSE,
  line_size = 1,
  legend_fs = 1,
  axis_fs = 1
)
}
\arguments{
\item{sig_list}{Output generated from profile_extract}

\item{color}{Manual colors for each bigWig. Default NULL.}

\item{condition}{Default. Condition associated with each bigWig. Lines will colord accordingly.}

\item{condition_colors}{Manual colors for each level in condition. Default NULL.}

\item{collapse_replicates}{Default FALSE. If TRUE and when \code{condition} is given, collapse signals samples belonging to same condition}

\item{plot_se}{Default FALSE. If TRUE plots standard error shading}

\item{line_size}{Default 1}

\item{legend_fs}{Legend font size. Default 1}

\item{axis_fs}{Axis font size. Default 1}
}
\description{
Draw a profile plot
}
