% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{profile_summarize}
\alias{profile_summarize}
\title{Summarize data for profile plots}
\usage{
profile_summarize(sig_list = NULL, stat = "mean", condition = NULL)
}
\arguments{
\item{sig_list}{Output generated from \code{profile_extract}}

\item{stat}{Default \code{mean}. Can be \code{mean}, \code{median}}

\item{condition}{column name with conditions in \code{colData}. If provided summarizes signals from samples belonging to same group or condition}
}
\description{
Summarize data for profile plots
}
