% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{read_coldata}
\alias{read_coldata}
\title{Prepares meta data table from bigWig files.
Output from this function is passed to all downstream functions.}
\usage{
read_coldata(
  bws = NULL,
  sample_names = NULL,
  build = "hg38",
  input_type = "bw"
)
}
\arguments{
\item{bws}{path to bigWig files}

\item{sample_names}{sample names for each input files. Optional. Default NULL - creates one from file names.}

\item{build}{Reference genome build. Default hg38}

\item{input_type}{Default \code{bw}. Can be \code{bw} or \code{peak}}
}
\description{
Prepares meta data table from bigWig files.
Output from this function is passed to all downstream functions.
}
\examples{
bigWigs = system.file("extdata", "bw", package = "trackplot") |> list.files(pattern = "\\\\.bw$", full.names = TRUE) 
cd = read_coldata(bws = bigWigs, build = "hg19")
beds = system.file("extdata", "narrowpeak", package = "trackplot") |> list.files(pattern = "\\\\.bed$", full.names = TRUE) 
cd_bed = read_coldata(bws = beds, input_type = "peak", build = "hg19")
}
