% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{track_plot}
\alias{track_plot}
\title{Generate IGV style locus tracks with ease}
\usage{
track_plot(
  summary_list = NULL,
  draw_gene_track = TRUE,
  query_ucsc = TRUE,
  show_ideogram = FALSE,
  build = "hg19",
  col = "gray70",
  groupAutoScale = TRUE,
  groupScaleByCondition = FALSE,
  y_max = NULL,
  txname = NULL,
  genename = NULL,
  gene_track_height = 2,
  scale_track_height = 1,
  show_axis = TRUE,
  gene_fsize = 0.6,
  track_names = NULL,
  track_names_pos = 0,
  regions = NULL,
  region_width = 1,
  collapse_txs = FALSE,
  boxcol = "#192A561A",
  boxcolalpha = 0.2,
  gene_model = NULL,
  isGTF = FALSE
)
}
\arguments{
\item{summary_list}{Output from track_extract}

\item{draw_gene_track}{Default FALSE. If TRUE plots gene models overlapping with the queried region}

\item{query_ucsc}{Default FALSE. But switches to TRUE when `gene_model` is not given. Requires `mysql` installation.}

\item{show_ideogram}{Default TRUE. If TRUE plots ideogram of the target chromosome with query loci highlighted. Works only when `query_ucsc` is TRUE.}

\item{build}{Genome build. Default `hg19`}

\item{col}{Color for tracks. Default `#2f3640`. Multiple colors can be provided for each track}

\item{groupAutoScale}{Default TRUE}

\item{groupScaleByCondition}{Scale tracks by condition}

\item{y_max}{custom y axis limits for each track. Recycled if required.}

\item{txname}{transcript name to draw. Default NULL. Plots all transcripts overlapping with the queried region}

\item{genename}{gene name to draw. Default NULL. Plots all genes overlapping with the queried region}

\item{gene_track_height}{Default 2}

\item{scale_track_height}{Default 1}

\item{show_axis}{Default FALSE}

\item{gene_fsize}{Font size. Default 1}

\item{track_names}{Default NULL}

\item{track_names_pos}{Default 0 (corresponds to left corner)}

\item{regions}{genomic regions to highlight. A data.frame with at-least three columns containing chr, start and end positions.}

\item{collapse_txs}{Default FALSE. Whether to collapse all transcripts belonging to same gene into a unified gene model}

\item{boxcol}{color for highlighted region. Default "#192A561A"}

\item{boxcolalpha}{Default 0.5}

\item{gene_model}{File with gene models. Can be a gtf file or UCSC file format. If you have read them into R as a data.frame, that works as well. Default NULL, automatically fetches gene models from UCSC server}

\item{isGTF}{Default FALSE. Set to TRUE if the `gene_model` is a gtf file.}
}
\description{
Generate IGV style locus tracks with ease
}
