% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{track_summarize}
\alias{track_summarize}
\title{Summarize tracks per condition}
\usage{
track_summarize(summary_list = NULL, condition = NULL, stat = "mean")
}
\arguments{
\item{summary_list}{Output from track_extract. Required.}

\item{condition}{a column name in \code{coldata} containing sample conditions. Default NULL.}

\item{stat}{can be \verb{mean, median}, \code{max}, \code{min}. NAs are excluded.}
}
\description{
Summarize tracks per condition
}
