% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot for limma results from differential peak Analysis}
\usage{
volcano_plot(
  res = NULL,
  fdr = 0.1,
  upcol = "#d35400",
  downcol = "#1abc9c",
  alpha = 0.6,
  size = 0.8
)
}
\arguments{
\item{res}{Output from \code{\link{diffpeak}}}

\item{fdr}{FDR threshold. Default 0.1}

\item{upcol}{color for up-regulated. Default "#d35400"}

\item{downcol}{color for up-regulated. Default "#1abc9c"}

\item{alpha}{Default 0.6}

\item{size}{Point size. Default 0.8}
}
\description{
Volcano plot for limma results from differential peak Analysis
}
\details{
Takes output from \code{diffpeak} and draws volcano plot
}
