% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{extract_summary}
\alias{extract_summary}
\title{Extract area under the curve for every peak from from given bigWig files.}
\usage{
extract_summary(
  bigWigs = NULL,
  bed = NULL,
  binSize = 50,
  nthreads = 4,
  ucsc_assembly = NULL,
  ucsc_startFrom = "start",
  ucsc_up = 2500,
  ucsc_down = 2500,
  custom_names = NULL
)
}
\arguments{
\item{bigWigs}{bigWig files. Default NULL. Required.}

\item{bed}{bed file or a data.frame with first 3 column containing chromosome, star, end positions.}

\item{binSize}{bin size to extract signal. Default 50 (bps). Should be >1}

\item{nthreads}{Default 4}

\item{ucsc_assembly}{If \code{bed} file not provided, setting \code{ucsc_assembly} to ref genome build will fetch transcripts from UCSC genome browser. e.g; 'hg19'}

\item{ucsc_startFrom}{Default "center". Can be "center", "start" or "end"}

\item{ucsc_up}{extend upstream by this many bps from \code{startFrom}. Default 2500}

\item{ucsc_down}{extend downstream by this many bps from \code{startFrom}. Default 2500}

\item{custom_names}{Default NULL and Parses from the file names.}
}
\description{
Extract area under the curve for every peak from from given bigWig files.
}
