% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{pca_plot}
\alias{pca_plot}
\title{Draw a PCA plot}
\usage{
pca_plot(
  summary_list = NULL,
  top = 1000,
  color = NULL,
  condition = NULL,
  condition_colors = NULL,
  show_cree = TRUE
)
}
\arguments{
\item{summary_list}{output from extract_summary}

\item{top}{Top most variable peaks to consider for PCA. Default 1000}

\item{color}{Manual colors for each bigWig. Default NULL.}

\item{condition}{Default. Condition associated with each bigWig. Lines will colord accordingly.}

\item{condition_colors}{Manual colors for each level in condition. Default NULL.}

\item{show_cree}{If TRUE draws a cree plot. Default TRUE}
}
\description{
Draw a PCA plot
}
