% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{profile_extract}
\alias{profile_extract}
\title{Generate profile plots with ease}
\usage{
profile_extract(
  bigWigs = NULL,
  bed = NULL,
  binSize = 50,
  startFrom = "center",
  up = 2500,
  down = 2500,
  ucsc_assembly = NULL,
  nthreads = 4,
  custom_names = NULL
)
}
\arguments{
\item{bigWigs}{bigWig files. Default NULL. Required.}

\item{bed}{bed file or a data.frame with first 3 column containing chromosome, star, end positions.}

\item{binSize}{bin size to extract signal. Default 50 (bps). Should be >1}

\item{startFrom}{Default "center". Can be "center", "start" or "end"}

\item{up}{extend upstream by this many bps from `startFrom`. Default 2500}

\item{down}{extend downstream by this many bps from `startFrom`. Default 2500}

\item{ucsc_assembly}{If `bed` file not provided, setting `ucsc_assembly` to ref genome build will fetch transcripts from UCSC genome browser. e.g; 'hg19'}

\item{nthreads}{Default 4}

\item{custom_names}{Default NULL and Parses from the file names.}
}
\description{
Generate profile plots with ease
}
