% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{profile_plot}
\alias{profile_plot}
\title{Draw a profile plot}
\usage{
profile_plot(
  sig_list = NULL,
  color = NULL,
  line_size = 1,
  legend_fs = 1,
  axis_fs = 1,
  xlab = NA,
  ylab = NA
)
}
\arguments{
\item{sig_list}{Output generated from profile_summarize}

\item{color}{Manual colors for each bigWig. Default NULL.}

\item{line_size}{Default 1}

\item{legend_fs}{Legend font size. Default 1}

\item{axis_fs}{Axis font size. Default 1}

\item{xlab}{x axis label. Default NA}

\item{ylab}{y axis label. Default NA}
}
\description{
Draw a profile plot
}
