% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{read_coldata}
\alias{read_coldata}
\title{Prepares meta data table from bigWig files.
Output from this function is passed to all downstream functions.}
\usage{
read_coldata(
  coldata = NULL,
  files_idx = NULL,
  sampleNames_idx = NULL,
  build = "hg38",
  input_type = "bw"
)
}
\arguments{
\item{coldata}{a data.frame or a tsv file containing absolute path to bigWig files and other optional associated sample annotations}

\item{files_idx}{column index containing path to bigWig (or narrowPeak, broadPeak)files. Required.}

\item{sampleNames_idx}{column index containing sample names for corresponding bigWig files. Optional. Default NULL - creates one from bigWig files.}

\item{build}{Reference genome build. Default hg38}

\item{input_type}{Default \code{bw}. Can be \code{bw} or \code{peak}}
}
\description{
Prepares meta data table from bigWig files.
Output from this function is passed to all downstream functions.
}
\examples{
bigWigs = system.file("extdata", "bw", package = "trackplot") |> list.files(pattern = "\\\\.bw$", full.names = TRUE) 
coldf = data.frame(bigWigs)
cd = read_coldata(coldata = coldf, files_idx = 1)
beds = system.file("extdata", "narrowpeak", package = "trackplot") |> list.files(pattern = "\\\\.bed$", full.names = TRUE) 
coldf_bed = data.frame(beds)
cd_bed = read_coldata(coldata = coldf_bed, files_idx = 1)
}
