% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{track_extract}
\alias{track_extract}
\title{Extract bigWig track data for the given loci}
\usage{
track_extract(
  bigWigs = NULL,
  loci = NULL,
  binsize = 50,
  custom_names = NULL,
  nthreads = 1
)
}
\arguments{
\item{bigWigs}{bigWig files. Default NULL. Required.}

\item{loci}{target region to plot. Should be of format "chr:start-end". e.g; chr3:187715903-187752003 OR chr3:187,715,903-187,752,003}

\item{binsize}{bin size to extract signal. Default 50 (bps).}

\item{custom_names}{Default NULL and Parses from the file names.}

\item{nthreads}{Default 1. Number of threads to use.}
}
\description{
Extract bigWig track data for the given loci
}
