% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplot.R
\name{track_summarize}
\alias{track_summarize}
\title{Summarize tracks per condition}
\usage{
track_summarize(summary_list = NULL, condition = NULL, stat = "mean")
}
\arguments{
\item{summary_list}{Output from track_extract. Required.}

\item{condition}{Default NULL. Provide condition for each bigWig file}

\item{stat}{can be `mean, median`, `max`, `min`. NAs are excluded.}
}
\description{
Summarize tracks per condition
}
